+++ src/rulesets/oas3/index.json
@@ -96,6 +96,17 @@
       "tags": [
         "schema"
       ]
+    },
+    "$ref-siblings": {
+      "description": "$ref cannot be placed among any other property",
+      "message": "{{error}}",
+      "type": "validation",
+      "severity": 0,
+      "recommended": true,
+      "given": "$",
+      "then": {
+        "function": "refSiblings"
+      }
     }
   }
 }
